+++ metadata-ingestion/src/datahub/cli/delete_cli.py
@@ -10,7 +10,6 @@ from requests import sessions
 from tabulate import tabulate
 
 from datahub.cli import cli_utils
-from datahub.cli.cli_utils import guess_entity_type
 from datahub.emitter import rest_emitter
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
 from datahub.metadata.schema_classes import (
@@ -20,6 +19,7 @@ from datahub.metadata.schema_classes import (
 )
 from datahub.telemetry import telemetry
 from datahub.upgrade import upgrade
+from datahub.utilities.urns.urn import guess_entity_type
 
 logger = logging.getLogger(__name__)
