+++ metadata-ingestion/src/datahub/emitter/mcp_builder.py
@@ -1,6 +1,6 @@
 import hashlib
 import json
-from typing import Any, Iterable, List, Optional, TypeVar, Union
+from typing import Any, Iterable, List, Optional, TypeVar
 
 from pydantic.fields import Field
 from pydantic.main import BaseModel
@@ -16,6 +16,7 @@ from datahub.metadata.schema_classes import (
     ContainerClass,
     DomainsClass,
     GlobalTagsClass,
+    MetadataChangeEventClass,
     OwnerClass,
     OwnershipClass,
     OwnershipTypeClass,
@@ -23,6 +24,7 @@ from datahub.metadata.schema_classes import (
     TagAssociationClass,
     _Aspect,
 )
+from datahub.utilities.urns.urn import guess_entity_type
 
 
 class DatahubKey(BaseModel):
@@ -250,7 +252,7 @@ def add_dataset_to_container(
     # FIXME: Union requires two or more type arguments
     container_key: KeyType,
     dataset_urn: str,
-) -> Iterable[Union[MetadataWorkUnit]]:
+) -> Iterable[MetadataWorkUnit]:
     container_urn = make_container_urn(
         guid=container_key.guid(),
     )
@@ -280,3 +282,17 @@ def add_entity_to_container(
     )
     wu = MetadataWorkUnit(id=f"container-{container_urn}-to-{entity_urn}", mcp=mcp)
     yield wu
+
+
+def mcps_from_mce(
+    mce: MetadataChangeEventClass,
+) -> Iterable[MetadataChangeProposalWrapper]:
+    for aspect in mce.proposedSnapshot.aspects:
+        yield MetadataChangeProposalWrapper(
+            entityType=guess_entity_type(mce.proposedSnapshot.urn),
+            changeType=ChangeTypeClass.UPSERT,
+            entityUrn=mce.proposedSnapshot.urn,
+            auditHeader=mce.auditHeader,
+            aspect=aspect,
+            systemMetadata=mce.systemMetadata,
+        )
